Finishing up reverse custom url routing
In this change:
 - Add customurl fields to projects and users.
 - Updated services, fixtures and tests with the new field.
 - New get_custom_url method on projects.
 - Fixed some errors on the homepage.

+++ app/openfire/services/url.py
@@ -47,8 +47,16 @@ class CustomUrlService(RemoteService):
             raise remote.ApplicationError('Custom URL already exists.')
 
         target = ndb.Key(urlsafe=request.target)
+        target_obj = target.get()
+        if not target_obj:
+            # Target object not found.
+            raise remote.ApplicationError('put url requires a valid target')
+
         custom_url = CustomURL(key=url_key, slug=request.slug, target=target)
         custom_url.put()
+
+        target_obj.customurl = url_key
+        target_obj.put()
 
         return custom_url.to_message()
 
@@ -58,6 +66,15 @@ class CustomUrlService(RemoteService):
 
         ''' Remove a custom URL from a target, if it exists. '''
         url_key = ndb.Key(urlsafe=request.key)
+        url_obj = url_key.get()
+        if not url_obj:
+            raise remote.ApplicationError('no url found to delete')
+
+        target = url_obj.target.get()
+        if target:
+            target.customurl = None
+            target.put()
+
         url_key.delete()
         return Echo(message='custom url removed')
